commit 2:  profile module with its related classes and events added
1_ profileController.java is the main java class file holding
   ProfileService class
   profileRepository interface
   Profile record
   ProfileStatus enum
   ProfileType enum
   AddProfileRequest record
   UpdateProfileRequest record
   ProfileResponse Record
   ProfileAlreadyExistsException class
   ProfileAlreadyDisabledException class
   ProfileAlreadyEnabledException class

A hardcoded String which is a UUID of the Credential class (Authentication module or user) is defined in class which is used while sending request to service methods so that an authenticated user can only access his/her profile details.

Rest resource naming strategy tried to convey a REST API’s resource model.