commit 2:  product module with its related classes and events added
1_ productController.java is the main java class file for this module holding
   ProductService class
   ProductRepository interface
   Product record
   ProductStatus enum
   AddProductRequest record
   UpdateProductRequest record
   ProductResponse Record
   ProductAlreadyExistsException class
   ProductNotFoundException class
   ProductAlreadyDisabledException class
   ProductAlreadyEnabledException class

A hardcoded String which is a UUID of the Credential class (Authentication module or user) is defined in class which is used while sending request to service methods so that an authenticated user can only access his/her product details.

Rest resource naming strategy tried to convey a REST API’s resource model.