commit: catalog module added
Catalog module consists of two main entities namely Catalog and Category.

Before this commit I decided on setting Catalog to be the aggregate root in this module and the category to be the child, the main issue was that addition,update and deletion of a single category must have happened in the catalog (DDD), therefor update or deletion of a single category caused update of the catalog, deletion of all its categories and re-insert of all deleted categories plus the new one into the database just like below code:

update method in catalog record (aggregate root version)

Catalog updateCategory(UUID categoryId,Category category){
	 Category existsById = categories.stream()
	            .filter(c -> c.id().equals(categoryId))
			    .findFirst()
			    .orElseThrow(() -> new CategoryNotFoundException("Category with id " + categoryId + " not found"));

	    categories.remove(existsById);
	    categories.add(category);
	    return this;
}
 and then calling catalogRepository.save(catalog) in service method, in this approach catalog could keep version number of its column but not category due to the fact that catalog was aggregate root and, it was logical not to keep category version.

There was also a hybrid approach is such a way that I could do the remove in updateCategory() (to keep consistency) as above and then call categoryRepository.save(category) so that preventing deletion and re-insert of all categories into the database, however a new issue arose which was version mismatch, catalog could not keep the version of its columns but category could do, so I finally decided to forget all about aggregate root and its children as the real issue was choosing the wrong domain here, catalog could not be the aggregate root and keep the state of categories, also category was going to have recursive relationship to itself. So I decided to choose the third approach, each entity is its own aggregate root so that they can keep track of their state,version and no more extra code.

Classes and interfaces added are as bellow;

AddCategoryEvent
UpdateCategoryEvent
AddCatalogEvent
UpdateCatalogEvent
CatalogController with all its server,repository,dtos,mappers... all package private
CategoryController with all its server,repository,dtos,mappers... all package private
CategoryApi interface which is public to all other modules

Also, I update the sql schema as well as addition of catalog.http and category.http to test the controller endpoints.
implementation of methods related to addition,deletion and update of children categories to parent categories will be done soon.