commit: category entity endpoints refactored
In order for the categoryController methods to be REST-compliant a few endpoints uri refactored so that instead of sending catalogId through add and update endpoints uris I decided to send it in AddCategoryRequest and UpdateCategoryRequest dtos.

The Delete method refactored to return 204 status code instead of returning a list of existing categories. (after the category successfully deleted)