commit: Closure Table pattern introduced to category
Closure table pattern is used to store hierarchies, in fact closure table stores all transitive relationships between nodes in a separate table, it eliminates the need for recursive queries.

1_ The method names updated in CategoryController.

2_ Exception handling (throwing exceptions) enhanced with localization messages which placed in messages.properties file.

3_ The methods introduced in CategoryController are

1_addAncestor
2_addDescendant
3_update
4_findById
5_findByName
6_findBySlug
7_activate (a single category)
8_deactivate (a single category)
9_findAll
10_findAllAncestors
11_findAllDescendants
12_activateOrDeactivateParentAndChildrenWithoutVersioning (without handling version and updatedAt functionality)
13_activateOrDeactivateParentAndChildrenWithVersioning (with handling version and updatedAt functionality)
14_deleteCategory

also CategoryErrorCode a new enum added to the category file which is used in combination with custom exceptions to provide more informative error messages.