commit: A new method called findAllWithRootCategories add to Catalog
A new method called findAllWithRootCategories added to the Catalog class which returns all the existing catalogs along with their categories that are root categories (categories with no children)

1_Three dto records added to the class
a_CatalogDTO
b_CategoryDto
c_CatalogWithRootCategory

2_The findAllWithRootCategories() method in CatalogRepository:

The custom sql query used for findAllWithRootCategories method selects attributes from catalog and category tables , it performs two left joins, one to categories table and the other one to category_closure table where depth = 1 and parent_id is null and the method returns a List of CatalogWithRootCategory record dto, and it takes limit and offset (custom pageable).

3_The findAllWithRootCategories() method in CatalogService:
This method takes two parameters namely page and size, at first it calculates the int offset, it then calls the findAllWithRootCategories() method in the repository, then creates a Map<UUID, CatalogDto> and next performs a for loop.

The computeIfAbsent is used to create a catalog only once ,even though there may be multiple categories for the corresponding catalog, the computeIfAbsent checks if the catalogMap already has a catalogDto for the current row.catalog_id() , if it does not have the catalogMap it creates a new CatalogDto otherwise it does nothing, and finally it adds the category (if exists) to the catalog's rootCategories.

Note:
@EnableSpringDataWebSupport(pageSerializationMode = EnableSpringDataWebSupport.PageSerializationMode.VIA_DTO)

4_The findAllWithRootCategories() method in CatalogController
This method returns Page<CatalogDto> and it takes page and size as its attributes.

5_sql schema updated - two column created_at and updated_at added to the catalogs table.

6_message properties updated.
7_catalog.http updated
8_category.http updated