commit: findAllWithRootCategories refactored
In the previous commit I added a method to the catalog class to select all existing catalogs including their root categories which were root categories (categories with no parent), this method also provided pagination, the reason behind implementing it was that to perform this scenario (selecting catalogs....) a custom query was required to perform multiple joins to category and category-closure tables, as far as I knew spring data jdbc does not provide pagination for custom queries.

The issue with the previous method was that it was returned all existing catalogs with their corresponding root categories but the total number of returned element was not matching with the data in the database.(instead of just counting catalogs it used to count each catalog plus its categories).

Refactoring steps:
What I did was to first define two separate select query in the repository
1- select all existing catalogs with pagination functionality.
2- select all root categories of each catalogId (the method in the repository takes a List of catalogIds)

3- Keep the 3 dto records:
    catalogDto
    CategoryDto
    CatalogWithRootCategory

4_update the findAllWithRootCategories(int page, int size) in the service class
    a_calculate the offset
    b_find all catalogs
    c_if it returned empty then returned an empty PageImpl
    d_fetch the catalogIds from catalogs
    e_ map catalogIds to CatalogDto
    f_ attach categories to CatalogDto
    g_ calculate total number of catalogs
    h- return the paged dtos

I also added CorsConfig as the React frontend added to the project, but for now it is ignored from committing to GitHub.