commit: a new method called findByIdWithRootCategories added to catalog
This method is defined to display a selected catalog and its root categories.

Repository layer:
The method in the repository takes a catalogId and returns a list of CatalogWithRootCategory dto. The custom query of this method selects the catalog's detail as well as its categories' data through two joins to categories and category_closure tables .

Service layer:
The method in service layer takes the uuid of a catalog and returns a CatalogDto object. It first checks if the catalog exists in the database, then calls the findByIdAndWithRootCategories in the repository and returns a list of CatalogWithRootCategory, it then creates a catalogDto Object with and initializes an empty list for rootCategories, and finally through a for loop adds rootCategories to the catalogDto only if the category_id is not null and then returns the catalogDto.

Controller layer:
The controller method endpoint simply is "/api/catalogs/{id}/with-root-categories".

A few updates also added to the project that are subject to change.