commit: A new method called findAllRootCategories added to Category
Category updated with a new method called findAllRootCategories which fetches all the categories without parent with depth 1, this method is used when the admin visits the category page in frontend.

Repository layer
The custom query selects all the required attributes corresponding to root categories and , in a sub query it checks whether if the selected category is a child with depth 1 in order to ensure that the selected categories are not children of other categories.

The method returns a List of Categories which then will be mapped to a custom dto called RootCategoryResponse in the service layer.

Service layer
The method in the service layer has the same as the method in the repository, it takes two parameter page and size, in the method body it first calculated the offset, fetches all the root categories as well as counting the number of all root categories (using a method called countRootCategories which is defined in the repository layer). finally using the mapper it converts List of categories to a list of RootCategoryResponse and creates a PageImpl using rootCategoryResponseList , total and   PageRequest.of(page,size).

Controller layer
In the controller layer the method return a  Page<RootCategoryResponse> and takes page and size as its parameters, the uri to this method is "/api/categories/root".