commit: Category refactored with basic mapping and caching functionality
In order to simplify the method called findACategoryWithSubCategoriesById() in the service layer two mapping methods defined in CategoryMapper interface.

1_ CategoryDtoWithSubs toCategoryDtoWithSubs(Category category);
2_ CategoryDtoWithSubs toCategoryDtoWithSubs(Category category, List<CategoryDtoWithSubs> subCategories);

in addition to defining these two mapping methods caching annotation added to findACategoryWithSubCategoriesById(UUID id) in the service layer, also cacheEvict enabled for this method on addDescendant,update and delete method.

 @Caching(evict = {
        @CacheEvict(value = "catalogs", allEntries = true),
        @CacheEvict(value = "categories", allEntries = true)
  })