commit: ImageUrl attribute added to catalog part1
In order for every catalog to have a displaying image, imageUrl attribute added to Catalog record, in addition to that a new mapping method called mapCatalogWithImage added to the CatalogMapper.

Service layer:

A service method called uploadImage and two helper methods called prepareImageUpload and imageExtension added to the service class.