commit: custom exceptions revised in Advertisement module
In order to keep consistency across all custom exceptions in the entire project, custom exception classes in the advertisement module refactored, now exceptions in this module show errorCodes related to the type of exception thrown. The enum that contains error types is called AdvertisementErrorCode:

public enum AdvertisementErrorCode {
	ADVERTISEMENT_NOT_FOUND,
	ADVERTISEMENT_ALREADY_EXISTS,
	ADVERTISEMENT_ALREADY_ACTIVATED,
	ADVERTISEMENT_ALREADY_DEACTIVATED
}

The AdvertisementExceptionHandler which is a central exception handler to advertisement now handle 4 types of exceptions (local exceptions to the advertisement module).