commit: A method called uploadMultipleImages defined in AdvertisementImageController | not completed 1
A service method called uploadMultipleImages defined in AdvertisementImageController that us responsible to receive a list of MultipartFile as well as a list of AddImageRequest (in other words Metadata for each image) and store each image in the filesystem and their metadata in the database.

The method is quite lengthy, and it requires multiple drastic refactoring and revisions.