commit: single and multiple image upload enhanced in AdvertisementImageController
Now the controller here exposed two method to upload single and multiple image upload, each image including its metadata will be stored in the database.

In the single image upload method
---------------------------------
in the service layer, it checks if advertisement exists at all , then in check if already advertisement has stored metadata images which means it has already saved images in the db, I already restricted the maximum number of images that can be uploaded for an advertisement which is 5.

The method create a main directory with the advertisementId, stored the image in the filesystem , it then maps AddImageRequest to AdvertisementImage and append a url of the stored images to the AdvertisementImage and save its metadata in the db and returns a string url of the stored image.

In the multiple image upload method
---------------------------------
The method almost does the same steps as the previous method but this time for more than one image also A removedInconsistentImages is added to the catch block to removed stored images if any database error happens during execution.

The method yet has lots of work to be completed.