commit: saveImageMetadata in ImageMetadataController file revised | with no bugs
saveImageMetadata() method is s service method in the ImageMetadataService, it is obviously responsible to save the metadata of each image file that it gets from the controller, also this method returns a list of ImageMetadataResponses.

A new method called uploadMultipleImages introduced to the ImageMetadataService service class. This method combine all the necessary calls to perform image upload successfully as follows:

1_saveImageMetadata(metadataList, images);

    This method checks null safety , checks the Db for data consistency, checks if the number of images do not exceed from the MAX_IMAGES_PER_ADVERTISEMENT which is 5, if any image in the Db is set to main then it does not normalize the new list of ImageMetadata, and if there is no main image in the db then it normalize the current list of ImageMetadata.The method used for the mentioned scenario are
    saveWithOneIsSetToMain(requests, files);
    savedMetadata = saveWithoutMain(requests, files);

    In addition to the above functionality it also saves the ImageMetadata with
       PENDING: status as at this stage the image is not uploaded into the filesystem
       URL:  with null value(or it can be a default value), it is null as the url is not yet created.
       VERSION
       ID
       INSERTED_AT

2_createMainDirectoryIfNotExists(imageMetadataResponses,imagePathProperties.getAdvertisementImagePath()); which creates a base directory for each advertisementId;

3_ The method here is called to read the image file completely from the request for consistency (prevents image data corruption):
	List<byte[]> fileContents = images.stream().map(file -> {try {return file.getBytes();} catch (IOException e) {throw new UncheckedIOException(e);}}).toList();

4_Previously I used to create a baseUrl inside prepareImageUpload method of ImageUploadService class for each image by calling
  String baseUrl = ServletUriComponentsBuilder.fromCurrentContextPath().path("/api/advertisements/images/" + imageName).toUriString();

and then call the prepareImageUpload method inside the async method , the issue was that the baseUrl string for the second call to the prepareImageUpload method was not created at all due to the thread mismatch between the main thread and the async thread, so that I created the base url in the uploadMultipleImages once and reused its string by passing it to the storeImagesIntoFileSystemAsync.

5_ The async method :

this.imageUploadService.storeImagesIntoFileSystemAsync(imageMetadataResponses,fileContents,imageBasePath,baseUrl);

This method is the vital method here as it  prevents the main thread from being responsible for uploading images which is a heavy task.
Inside the storeImagesIntoFileSystemAsync method, after each successful image upload, and event that is UpdateImageMetadataEvent get published which takes the id of the imageMetadata and the url of the image being stored in the file system.The event listener responsible to handle this event is called ImageMetadataUpdateEventHandler which takes the event data and checks if the corresponding ImageMetadata with the provided id exists in the db , if yes it updates the ImageMetadata with the provided url and succeed status.