commit: synchronized image upload functionality added to Catalog
Synchronized image upload functionality added to Catalog, for this a new service called
CatalogImageUploadService added to the Catalog codebase that contains a number of handy
methods including:

A_createMainDirectoryIfNotExists(UUID id, String path)
B_imageExtension(String imageName)
C_removeExtension(String filename)
D_preloadImage(MultipartFile file)
E_storeImagesIntoFileSystemAsync(UUID imageId, byte[] imageFile,Path imageBasePath,String baseUrl) => the sync method

in addition to defining the CatalogImageUploadService I needed to define an event-listener,an even record in order to publish and listen to the event when the image file gets saved in the filesystem, and it is time to update the imageUrl  in the Catalog record in the DB.

All the necessary log messages for the process of uploading the image are also included in the codebase.

A new config class called CatalogAsyncConfig to config thread pool for asynchronous task defined in the catalog module.