commit: Added PATCH request in CatalogController.
In order to make partial updates Patch request added in CatalogController.

A- A PatchCatalogRequest that is a dto record with non-null fields added (except id).
B- A new mapping method called mapPatchRequestToCatalog added to CatalogMapper
C- A controller method called patch added to the controller with @PatchMapping annotation
D- A service method called patch added to the service

The service method checks the existence of the catalog by id, if it exists it then check if the new data sent from the frontend is equal with the existing ones, if they are not equal it checks for duplicates (both for name and slug), if there were no duplicates it then maps the request to the existing catalog and it the saves the newly entered data.

Since I used records to manipulate data and records are immutable, the mapping process is a bit different from mutable data, every time I call to map data from one type to another type, a new record object gets created therefor, in order to keep the existing fields (untouched data) in the DB and combine them with the changed fields (updated data, newly entered data) the mapping methods creates a new object, so being cautious about null values in the newly created object and saving null values in the DB is vital.

For instance if imageUrl already exists in the db for a record in order to prevent it from being updated is to take it from the sources and then save the new object:

    @Mapping(target = "id", source = "catalog.id")
	@Mapping(target = "version", source = "catalog.version")
	@Mapping(target = "name", expression = "java(request.name() != null ? request.name() : catalog.name())")
	@Mapping(target = "slug", expression = "java(request.slug() != null ? request.slug() : catalog.slug())")
	@Mapping(target = "description", expression = "java(request.description() != null ? request.description() : catalog.description())")
	@Mapping(target = "createdAt", source = "catalog.createdAt")
	@Mapping(target = "updatedAt", expression = "java(LocalDateTime.now())")
	@Mapping(target = "imageUrl", source = "catalog.imageUrl")
	Catalog mapPatchRequestToCatalog(PatchCatalogRequest request,Catalog catalog);