commit: Added PATCH request in CategoryController.
In order to make partial updates Patch request added in CategoryController.

A- A PatchCategoryRequest that is a dto record with non-null fields added (except id).
B- A new mapping method called mapPatchRequestToCategory added to CategoryMapper
C- A controller method called patch added to the controller with @PatchMapping annotation
D- A service method called patch added to the service

The service method checks the existence of the category by id, if it exists it then check if the new data sent from the frontend is equal with the existing ones, if they are not equal it checks for duplicates (both for name and slug), if there were no duplicates it then maps the request to the existing category and it the saves the newly entered data.