fix (authentication module): set the findByUserIdAndDeviceIdHash to return Optional instead of null
Made the findByUserIdAndDeviceIdHash method instead of null returns Optional, , this method is called in saveOrUpdate method to check whether if device already belong to the same user or not.

Now call to the findByUserIdAndDeviceIdHash method in saveOrUpdate return Optional of type deviceBinding , if it is empty it then saves the device otherwise it updates the device, also the if-else part removed from the saveOrUpdate.